requirements.txt
tensorflow-macos
pandas 
numpy
scikit-learn
tensorboard
matplotlib
streamlit